arreglo error de perdida de datos por conversion de int a float de Tusca

+++ trunk/randomgame/src/Servidor/modelo/TerrainProcessor.cpp
@@ -125,8 +125,8 @@ b2Vec2 TerrainProcessor::transformBmpToBox2D(b2Vec2 vertex, int height, int widt
 	{
 		ParserYaml* aParser = ParserYaml::getInstance();
 		b2Vec2 nuevo;
-		nuevo.y =((-1*vertex.x)+height-1)*(atoi((aParser->getEscenarioAltoU()).c_str() ) / height);
-		nuevo.x = (vertex.y)*(atoi((aParser->getEscenarioAnchoU()).c_str() ) / width);
+		nuevo.y =((-1*vertex.x)+height-1)*(atoi((aParser->getEscenarioAltoU()).c_str() ) / (float)height);
+		nuevo.x = (vertex.y)*(atoi((aParser->getEscenarioAnchoU()).c_str() ) / (float)width);
 		return nuevo;
 	}
